% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_mapping.R
\name{create_brms_family}
\alias{create_brms_family}
\title{Map glmmTMB Family to brms Family}
\usage{
create_brms_family(family_config, link = NULL)
}
\arguments{
\item{family_config}{Family configuration from get_family_config}

\item{link}{Optional link function override}
}
\value{
A brms family object
}
\description{
Creates the appropriate brms family object based on family configuration.
}
