% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_mapping.R
\name{create_glmmTMB_family}
\alias{create_glmmTMB_family}
\title{Create glmmTMB Family}
\usage{
create_glmmTMB_family(family_config)
}
\arguments{
\item{family_config}{Family configuration from get_family_config}
}
\value{
A glmmTMB family object
}
\description{
Creates the appropriate glmmTMB family object.
}
