% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_fitting.R
\name{fit_bayesian_zibeta}
\alias{fit_bayesian_zibeta}
\title{Fit Bayesian Zero-Inflated Beta Mixed Model (Deprecated)}
\usage{
fit_bayesian_zibeta(...)
}
\arguments{
\item{...}{Arguments passed to \code{fit_bayesian_model}}
}
\description{
This function is deprecated. Use \code{fit_bayesian_model} instead.
}
