% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequentist_fitting.R
\name{fit_frequentist_zibeta}
\alias{fit_frequentist_zibeta}
\title{Fit Frequentist Zero-Inflated Beta Mixed Model (Deprecated)}
\usage{
fit_frequentist_zibeta(...)
}
\arguments{
\item{...}{Arguments passed to \code{fit_frequentist_model}}
}
\description{
This function is deprecated. Use \code{fit_frequentist_model} instead.
}
