% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_fitting.R
\name{posterior_predict_model}
\alias{posterior_predict_model}
\title{Posterior Predictive Samples from Bayesian Model}
\usage{
posterior_predict_model(model, newdata, ndraws = 100, ...)
}
\arguments{
\item{model}{Fitted Bayesian model object.}

\item{newdata}{Data frame for predictions.}

\item{ndraws}{Number of posterior draws. Default is 100.}

\item{...}{Additional arguments.}
}
\value{
Matrix of posterior predictive samples (ndraws × nrow(newdata)).
}
\description{
Generate posterior predictive samples for new data.
}
