% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_fitting.R
\name{posterior_predict_zibeta}
\alias{posterior_predict_zibeta}
\title{Posterior Predictive Samples (Deprecated)}
\usage{
posterior_predict_zibeta(...)
}
\arguments{
\item{...}{Arguments passed to posterior_predict_model}
}
\description{
Posterior Predictive Samples (Deprecated)
}
