% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_prediction.R
\name{print.conformal_prediction}
\alias{print.conformal_prediction}
\title{Print Method for Conformal Prediction Results}
\usage{
\method{print}{conformal_prediction}(x, ...)
}
\arguments{
\item{x}{Conformal prediction results object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for Conformal Prediction Results
}
