% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{print.workflow_results}
\alias{print.workflow_results}
\title{Print Method for Workflow Results}
\usage{
\method{print}{workflow_results}(x, ...)
}
\arguments{
\item{x}{Workflow results object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for Workflow Results
}
