% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{run_zibeta_workflow}
\alias{run_zibeta_workflow}
\title{Run Zero-Inflated Beta Workflow (Deprecated)}
\usage{
run_zibeta_workflow(...)
}
\arguments{
\item{...}{Arguments passed to run_workflow}
}
\description{
Run Zero-Inflated Beta Workflow (Deprecated)
}
